State view refactor
Refactor all the screens with different compose views
