+++ .idea/misc.xml
@@ -136,6 +136,8 @@
         <entry key="../../../../../layout/compose-model-1652451826885.xml" value="0.20185185185185187" />
         <entry key="../../../../../layout/compose-model-1652456419021.xml" value="0.20185185185185187" />
         <entry key="../../../../../layout/compose-model-1652458948210.xml" value="0.20185185185185187" />
+        <entry key="../../../../../layout/compose-model-1652527097349.xml" value="0.1787037037037037" />
+        <entry key="../../../../../layout/compose-model-1652527202129.xml" value="0.1787037037037037" />
         <entry key="app/src/main/res/drawable-v24/ic_launcher_foreground.xml" value="0.131" />
         <entry key="app/src/main/res/drawable/bg_splash.xml" value="0.1" />
         <entry key="app/src/main/res/drawable/bg_splash_12.xml" value="0.1485" />
