+++ app/build.gradle
@@ -67,9 +67,6 @@ dependencies {
     androidTestImplementation "androidx.compose.ui:ui-test-junit4:$compose_version"
     debugImplementation "androidx.compose.ui:ui-tooling:$compose_version"
 
-    //Constraint
-    implementation("androidx.constraintlayout:constraintlayout-compose:1.0.0")
-
     //Navigation
     implementation("androidx.navigation:navigation-compose:2.4.2")
