+++ app/src/main/java/com/mightyhedgehog/doplanner/presentation/screen/calendar/CalendarUpdateHandler.kt
@@ -2,9 +2,9 @@ package com.mightyhedgehog.doplanner.presentation.screen.calendar
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.mightyhedgehog.doplanner.presentation.handler.SimpleHandler
+import com.mightyhedgehog.doplanner.app.core.UpdateHandler
 
-class CalendarUpdateHandler: SimpleHandler<Unit> {
+class CalendarUpdateHandler: UpdateHandler<Unit> {
     override val data: LiveData<Unit> = MutableLiveData()
 
     override fun update(value: Unit) {
