+++ app/src/main/java/com/mightyhedgehog/doplanner/presentation/screen/daily/DailyUpdateHandler.kt
@@ -2,9 +2,9 @@ package com.mightyhedgehog.doplanner.presentation.screen.daily
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.mightyhedgehog.doplanner.presentation.handler.SimpleHandler
+import com.mightyhedgehog.doplanner.app.core.UpdateHandler
 
-class DailyUpdateHandler : SimpleHandler<Unit> {
+class DailyUpdateHandler : UpdateHandler<Unit> {
     override val data: LiveData<Unit> = MutableLiveData()
 
     override fun update(value: Unit) {
